+++ frontend/src/common/components/AppDrop.vue
@@ -19,6 +19,7 @@ if (!dataTransfer) {
     return;
 }
 const payload = dataTransfer.getData(DATA_TRANSFER_PAYLOAD);
+console.log(payload);
 if (payload) {
     const transferData = JSON.parse(dataTransfer.getData(DATA_TRANSFER_PAYLOAD));
     emit('drop', transferData);
