+++ frontend/src/modules/constructor/ContentDiameter.vue
@@ -1,0 +1,161 @@
+<template>
+    <div class="content__diameter">
+          <div class="sheet">
+            <h2 class="title title--small sheet__title">Выберите размер</h2>
+
+            <div class="sheet__content diameter">
+              <template v-for="size in sizes" :key="size.id">
+                <label
+                  class="diameter__input"
+                  :class="`diameter__input--${getDiameterValue(size.name)}`"
+                  @click="emit('setDiameter', getDiameterValue(size.name))"
+                >
+                  <input
+                    type="radio"
+                    name="diameter"
+                    :value="getDiameterValue(size.name)"
+                    class="visually-hidden"
+                    :checked='props.diameter === getDiameterValue(size.name)'
+                  />
+                  <span>{{ size.name }}</span>
+                </label>
+              </template>
+            </div>
+          </div>
+        </div>
+</template>
+
+<script setup>
+import sizes from "../../mocks/sizes.json";
+
+const props = defineProps({
+  diameter: {
+    type: String,
+    required: true,
+    default: 'small',
+  }
+})
+
+const emit = defineEmits(["setDiameter"]);
+
+function getDiameterValue(diameter) {
+  let diameterValue = "";
+  if (diameter === "23 см") {
+    diameterValue = "small";
+  } else if (diameter === "32 см") {
+    diameterValue = "normal";
+  } else {
+    diameterValue = "big";
+  }
+  return diameterValue;
+}
+</script>
+
+<style lang="scss" scoped>
+@import "@/assets/scss/app.scss";
+// title
+.title {
+  box-sizing: border-box;
+  width: 100%;
+  margin: 0;
+  color: $black;
+  &--big {
+    @include b-s36-h42;
+  }
+  &--small {
+    @include b-s18-h21;
+  }
+}
+// visually-hidden
+.visually-hidden {
+  position: absolute;
+  overflow: hidden;
+  clip: rect(0 0 0 0);
+  width: 1px;
+  height: 1px;
+  margin: -1px;
+  padding: 0;
+  white-space: nowrap;
+  border: 0;
+  clip-path: inset(100%);
+}
+// sheet
+.sheet {
+  padding-top: 15px;
+  border-radius: 8px;
+  background-color: $white;
+  box-shadow: $shadow-light;
+}
+.sheet__title {
+  padding-right: 18px;
+  padding-left: 18px;
+}
+.sheet__content {
+  display: flex;
+  align-items: center;
+  flex-wrap: wrap;
+  margin-top: 8px;
+  padding-top: 18px;
+  padding-right: 18px;
+  padding-left: 18px;
+  border-top: 1px solid rgba($green-500, 0.1);
+}
+// diameter
+.diameter__input {
+  margin-right: 8.7%;
+  margin-bottom: 20px;
+  padding-top: 7px;
+  padding-bottom: 6px;
+  cursor: pointer;
+  span {
+    @include r-s16-h19;
+    position: relative;
+    padding-left: 46px;
+    &::before {
+      @include p_center_v;
+      width: 36px;
+      height: 36px;
+      content: "";
+      transition: 0.3s;
+      border-radius: 50%;
+      background-color: $green-100;
+      background-image: url("@/assets/img/diameter.svg");
+      background-repeat: no-repeat;
+      background-position: center;
+    }
+  }
+  &:nth-child(3n) {
+    margin-right: 0;
+  }
+  &--small {
+    span::before {
+      background-size: 18px;
+    }
+  }
+  &--normal {
+    span::before {
+      background-size: 29px;
+    }
+  }
+  &--big {
+    span::before {
+      background-size: 100%;
+    }
+  }
+  &:hover {
+    span::before {
+      box-shadow: $shadow-regular;
+    }
+  }
+  input {
+    &:checked + span::before {
+      box-shadow: $shadow-large;
+    }
+  }
+}
+.content__diameter {
+  width: 373px;
+  margin-top: 15px;
+  margin-bottom: 15px;
+}
+</style>
