+++ frontend/src/modules/constructor/ContentPizza.vue
@@ -1,0 +1,277 @@
+<template>
+    <div class="content__pizza">
+        <label class="input">
+            <span class="visually-hidden">Название пиццы</span>
+            <input
+                type="text"
+                name="pizza_name"
+                placeholder="Введите название пиццы"
+            />
+        </label>
+        <app-drop @drop="moveFilling">
+            <div class="content__constructor">
+                <div class="pizza" :class='`pizza--foundation--${props.dough}-${props.sauce}`'>
+                    <div class="pizza__wrapper">
+                    <template v-for='filling in Object.getOwnPropertyNames(props.fillings)'>
+                        <div class="pizza__filling" :class='`pizza__filling--${filling}`'></div>
+                    </template>
+                    </div>
+                </div>
+            </div>
+        </app-drop>
+        
+
+        <div class="content__result">
+            <p>Итого: 0 ₽</p>
+            <button type="button" class="button" disabled>Готовьте!</button>
+        </div>
+    </div>
+</template>
+
+<script setup>
+import AppDrop from '../../common/components/AppDrop.vue';
+
+const props = defineProps({
+  dough:{
+    type: String,
+    required: true,
+  },
+  sauce:{
+    type: String,
+    required: true,
+  },
+  fillings:{
+    type: Object,
+    required: true,
+  },
+  moveFilling:{
+    type: Function,
+    required: true,
+  }
+})
+
+</script>
+
+<style lang="scss" scoped>
+@import "@/assets/scss/app.scss";
+// input
+.input {
+  display: block;
+  span {
+    @include r-s14-h16;
+    display: block;
+    margin-bottom: 4px;
+  }
+  input {
+    @include r-s16-h19;
+    display: block;
+    box-sizing: border-box;
+    width: 100%;
+    margin: 0;
+    padding: 8px 16px;
+    transition: 0.3s;
+    color: $black;
+    border: 1px solid $purple-400;
+    border-radius: 8px;
+    outline: none;
+    background-color: $white;
+    font-family: inherit;
+    &:focus {
+      border-color: $green-500;
+    }
+  }
+  &:hover {
+    input {
+      border-color: $black;
+    }
+  }
+  &--big-label {
+    display: flex;
+    align-items: center;
+    span {
+      @include b-s16-h19;
+      margin-right: 16px;
+      white-space: nowrap;
+    }
+  }
+}
+.content__pizza {
+  width: 373px;
+  margin-top: 15px;
+  margin-bottom: 15px;
+}
+.content__constructor {
+  width: 315px;
+  margin-top: 25px;
+  margin-right: auto;
+  margin-left: auto;
+}
+.content__result {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-top: 25px;
+  p {
+    font-size: 24px;
+    line-height: 28px;
+    margin: 0;
+  }
+  button {
+    margin-left: 12px;
+    padding: 16px 45px;
+  }
+}
+// visually-hidden
+.visually-hidden {
+  position: absolute;
+  overflow: hidden;
+  clip: rect(0 0 0 0);
+  width: 1px;
+  height: 1px;
+  margin: -1px;
+  padding: 0;
+  white-space: nowrap;
+  border: 0;
+  clip-path: inset(100%);
+}
+// pizza
+.pizza {
+  position: relative;
+  display: block;
+  box-sizing: border-box;
+  width: 100%;
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: 100%;
+  &--foundation--big-creamy {
+    background-image: url("@/assets/img/foundation/big-creamy.svg");
+  }
+  &--foundation--big-tomato {
+    background-image: url("@/assets/img/foundation/big-tomato.svg");
+  }
+  &--foundation--small-creamy {
+    background-image: url("@/assets/img/foundation/small-creamy.svg");
+  }
+  &--foundation--small-tomato {
+    background-image: url("@/assets/img/foundation/small-tomato.svg");
+  }
+}
+.pizza__wrapper {
+  width: 100%;
+  padding-bottom: 100%;
+}
+.pizza__filling {
+  $bl: &;
+  position: absolute;
+  top: 0;
+  left: 0;
+  display: block;
+  width: 100%;
+  height: 100%;
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: 100%;
+  &::before,
+  &::after {
+    display: none;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    content: "";
+    background-image: inherit;
+  }
+  &--second {
+    &::before {
+      display: block;
+      transform: rotate(45deg);
+    }
+  }
+  &--third {
+    &::before {
+      display: block;
+      transform: rotate(45deg);
+    }
+    &::after {
+      display: block;
+      transform: rotate(-45deg);
+    }
+  }
+  &--ananas,
+  &--ananas.pizza__filling--second::before,
+  &--ananas.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/ananas.svg");
+  }
+  &--bacon,
+  &--bacon.pizza__filling--second::before,
+  &--bacon.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/bacon.svg");
+  }
+  &--blue_cheese,
+  &--blue.pizza__filling--second::before,
+  &--blue.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/blue_cheese.svg");
+  }
+  &--cheddar,
+  &--cheddar.pizza__filling--second::before,
+  &--cheddar.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/cheddar.svg");
+  }
+  &--chile,
+  &--chile.pizza__filling--second::before,
+  &--chile.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/chile.svg");
+  }
+  &--ham,
+  &--ham.pizza__filling--second::before,
+  &--ham.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/ham.svg");
+  }
+  &--jalapeno,
+  &--jalapeno.pizza__filling--second::before,
+  &--jalapeno.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/jalapeno.svg");
+  }
+  &--mozzarella,
+  &--mozzarella.pizza__filling--second::before,
+  &--mozzarella.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/mozzarella.svg");
+  }
+  &--mushrooms,
+  &--mushrooms.pizza__filling--second::before,
+  &--mushrooms.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/mushrooms.svg");
+  }
+  &--olives,
+  &--olives.pizza__filling--second::before,
+  &--olives.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/olives.svg");
+  }
+  &--onion,
+  &--onion.pizza__filling--second::before,
+  &--onion.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/onion.svg");
+  }
+  &--parmesan,
+  &--parmesan.pizza__filling--second::before,
+  &--parmesan.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/parmesan.svg");
+  }
+  &--salami,
+  &---salami.pizza__filling--second::before,
+  &---salami.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/salami.svg");
+  }
+  &--salmon,
+  &--salmon.pizza__filling--second::before,
+  &--salmon.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/salmon.svg");
+  }
+  &--tomatoes,
+  &--tomatoes.pizza__filling--second::before,
+  &--tomatoes.pizza__filling--third::after {
+    background-image: url("@/assets/img/filling-big/tomatoes.svg");
+  }
+}
+</style>
