+++ frontend/src/views/HomeView.vue
@@ -4,222 +4,76 @@
       <div class="content__wrapper">
         <h1 class="title title--big">Конструктор пиццы</h1>
 
-        <div class="content__dough">
-          <div class="sheet">
-            <h2 class="title title--small sheet__title">Выберите тесто</h2>
+        <ContentDough :dough='dough' @setDough='setDough' />
 
-            <div class="sheet__content dough">
-              <template v-for="dough in doughs" :key="dough.id">
-                <label
-                  class="dough__input"
-                  :class="getDoughStyle(dough.name_eng)"
-                >
-                  <input
-                    type="radio"
-                    name="dought"
-                    value="light"
-                    class="visually-hidden"
-                    checked
-                  />
-                  <b>{{ dough.name }}</b>
-                  <span>{{ dough.description }}</span>
-                </label>
-              </template>
-            </div>
-          </div>
-        </div>
+        <ContentDiameter :diameter='diameter' @setDiameter='setDiameter' />
 
-        <div class="content__diameter">
-          <div class="sheet">
-            <h2 class="title title--small sheet__title">Выберите размер</h2>
+        <ContentIngredients :sauce='sauce' :setSauce='setSauce' :fillings='fillings' @setFillings='setFillings' @drop='moveFilling'/>
 
-            <div class="sheet__content diameter">
-              <template v-for="size in sizes" :key="size.id">
-                <label
-                  class="diameter__input"
-                  :class="getDiameterStyle(size.name)"
-                >
-                  <input
-                    type="radio"
-                    name="diameter"
-                    value="small"
-                    class="visually-hidden"
-                  />
-                  <span>{{ size.name }}</span>
-                </label>
-              </template>
-            </div>
-          </div>
-        </div>
-
-        <div class="content__ingredients">
-          <div class="sheet">
-            <h2 class="title title--small sheet__title">
-              Выберите ингредиенты
-            </h2>
-
-            <div class="sheet__content ingredients">
-              <div class="ingredients__sauce">
-                <p>Основной соус:</p>
-                <template v-for="sauce in sauces" :key="sauce.id">
-                  <label class="radio ingredients__input">
-                    <input type="radio" name="sauce" value="tomato" checked />
-                    <span>{{ sauce.name }}</span>
-                  </label>
-                </template>
-              </div>
-
-              <div class="ingredients__filling">
-                <p>Начинка:</p>
-                <ul class="ingredients__list">
-                  <template
-                    v-for="ingredient in ingredients"
-                    :key="ingredient.id"
-                  >
-                    <li class="ingredients__item">
-                      <span
-                        class="filling"
-                        :class="getFillingStyle(ingredient.name_eng)"
-                        >{{ ingredient.name }}</span
-                      >
-
-                      <div class="counter counter--orange ingredients__counter">
-                        <button
-                          type="button"
-                          class="counter__button counter__button--minus"
-                          disabled
-                        >
-                          <span class="visually-hidden">Меньше</span>
-                        </button>
-                        <input
-                          type="text"
-                          name="counter"
-                          class="counter__input"
-                          value="0"
-                        />
-                        <button
-                          type="button"
-                          class="counter__button counter__button--plus"
-                        >
-                          <span class="visually-hidden">Больше</span>
-                        </button>
-                      </div>
-                    </li>
-                  </template>
-                </ul>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div class="content__pizza">
-          <label class="input">
-            <span class="visually-hidden">Название пиццы</span>
-            <input
-              type="text"
-              name="pizza_name"
-              placeholder="Введите название пиццы"
-            />
-          </label>
-
-          <div class="content__constructor">
-            <div class="pizza pizza--foundation--big-tomato">
-              <div class="pizza__wrapper">
-                <div class="pizza__filling pizza__filling--ananas"></div>
-                <div class="pizza__filling pizza__filling--bacon"></div>
-                <div class="pizza__filling pizza__filling--cheddar"></div>
-              </div>
-            </div>
-          </div>
-
-          <div class="content__result">
-            <p>Итого: 0 ₽</p>
-            <button type="button" class="button" disabled>Готовьте!</button>
-          </div>
-        </div>
+        <ContentPizza :pizzaState='pizzaState' :dough='dough' :sauce='sauce' :fillings='fillings' :moveFilling='moveFilling'/>
       </div>
     </form>
   </main>
 </template>
 
 <script setup>
-import doughs from "../mocks/dough.json";
-import ingredients from "../mocks/ingredients.json";
-import sauces from "../mocks/sauces.json";
-import sizes from "../mocks/sizes.json";
+import { reactive, ref } from 'vue'
+import ContentDough from '../modules/constructor/ContentDough.vue';
+import ContentDiameter from '../modules/constructor/ContentDiameter.vue';
+import ContentIngredients from '../modules/constructor/ContentIngredients.vue';
+import ContentPizza from '../modules/constructor/ContentPizza.vue';
 
-function getFillingStyle(ingredient_name) {
-  return `filling--${ingredient_name}`;
-}
-
-function getDiameterStyle(diameter) {
-  let styleName = "";
-  if (diameter === "23 см") {
-    styleName = "small";
-  } else if (diameter === "32 см") {
-    styleName = "normal";
-  } else {
-    styleName = "big";
+const dough = ref('small');
+const setDough = (value) => {
+    dough.value = value
   }
-  return `diameter__input--${styleName}`;
+
+const diameter = ref('small');
+const setDiameter = (value) => {
+    diameter.value = value
+  }
+
+const sauce = ref('tomato');
+const setSauce = (value) => {
+    sauce.value = value
+  }
+
+const fillings = ref({
+    bacon: 1,
+    ham: 1,
+    cheddar: 1,
+  });
+const setFillings = (value) => {
+    fillings.value = value
+  }
+
+  function moveFilling(fillingName){
+    console.log('moveFilling');
+    if(fillings.hasOwnProperty(fillingName)){
+        const fillingNewValue = fillings[fillingName] + 1;
+        fillings.value = {...fillings.value, [fillingName]: fillingNewValue};
+    }
+    else{
+        fillings.value = {...fillings.value, [fillingName]: 1};
+    }
 }
 
-function getDoughStyle(dough_name) {
-  return `dough__input--${dough_name}`;
-}
+const pizzaState = reactive({
+  name: '',
+  dough: 'small',
+  diameter: 'big',
+  sauce: 'tomato',
+  fillings: {
+    bacon: 1,
+    ham: 1,
+    cheddar: 1,
+  }
+});
+
 </script>
 
 <style lang="scss" scoped>
 @import "@/assets/scss/app.scss";
-// radio
-.radio {
-  cursor: pointer;
-  span {
-    @include r-s16-h19;
-    position: relative;
-    padding-left: 28px;
-    &:before {
-      @include p_center-v;
-      display: block;
-      box-sizing: border-box;
-      width: 20px;
-      height: 20px;
-      content: "";
-      transition: 0.3s;
-      border: 1px solid $purple-400;
-      border-radius: 50%;
-      background-color: $white;
-    }
-  }
-  &:hover {
-    input:not(:checked):not(:disabled) + span {
-      &:before {
-        border-color: $purple-800;
-      }
-    }
-  }
-  input {
-    display: none;
-    &:checked + span {
-      &:before {
-        border: 6px solid $green-500;
-      }
-    }
-    &:disabled {
-      & + span {
-        &:before {
-          border-color: $purple-400;
-          background-color: $silver-200;
-        }
-      }
-      &:checked + span {
-        &:before {
-          border: 6px solid $purple-400;
-        }
-      }
-    }
-  }
-}
 // input
 .input {
   display: block;
@@ -288,99 +142,7 @@ function getDoughStyle(dough_name) {
   padding-bottom: 30px;
   padding-left: 2.12%;
 }
-.content__dough {
-  width: 527px;
-  margin-top: 15px;
-  margin-right: auto;
-  margin-bottom: 15px;
-}
-.content__diameter {
-  width: 373px;
-  margin-top: 15px;
-  margin-bottom: 15px;
-}
-.content__ingredients {
-  width: 527px;
-  margin-top: 15px;
-  margin-right: auto;
-  margin-bottom: 15px;
-}
-.content__pizza {
-  width: 373px;
-  margin-top: 15px;
-  margin-bottom: 15px;
-}
-.content__constructor {
-  width: 315px;
-  margin-top: 25px;
-  margin-right: auto;
-  margin-left: auto;
-}
-.content__result {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-top: 25px;
-  p {
-    font-size: 24px;
-    line-height: 28px;
-    margin: 0;
-  }
-  button {
-    margin-left: 12px;
-    padding: 16px 45px;
-  }
-}
-// dough
-.dough__input {
-  position: relative;
-  margin-right: 8%;
-  margin-bottom: 20px;
-  padding-left: 50px;
-  cursor: pointer;
-  b {
-    @include r-s16-h19;
-    &::before {
-      @include p_center-v;
-      width: 36px;
-      height: 36px;
-      content: "";
-      transition: 0.3s;
-      border-radius: 50%;
-      background-repeat: no-repeat;
-      background-position: center;
-      background-size: cover;
-    }
-  }
-  span {
-    @include l-s11-h13;
-    display: block;
-  }
-  &--light {
-    b {
-      &::before {
-        background-image: url("@/assets/img/dough-light.svg");
-      }
-    }
-  }
-  &--large {
-    b {
-      &::before {
-        background-image: url("@/assets/img/dough-large.svg");
-      }
-    }
-  }
-  &:hover {
-    b::before {
-      box-shadow: $shadow-regular;
-    }
-  }
-  input {
-    &:checked + b::before {
-      box-shadow: $shadow-large;
-    }
-  }
-}
+
 // sheet
 .sheet {
   padding-top: 15px;
@@ -402,435 +164,5 @@ function getDoughStyle(dough_name) {
   padding-left: 18px;
   border-top: 1px solid rgba($green-500, 0.1);
 }
-// ingredients
-.ingredients__sauce {
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  width: 100%;
-  margin-bottom: 14px;
-  p {
-    @include r-s16-h19;
-    margin-top: 0;
-    margin-right: 16px;
-    margin-bottom: 10px;
-  }
-}
-.ingredients__input {
-  margin-right: 24px;
-  margin-bottom: 10px;
-}
-.ingredients__filling {
-  width: 100%;
-  p {
-    @include r-s16-h19;
-    margin-top: 0;
-    margin-bottom: 16px;
-  }
-}
-.ingredients__list {
-  @include clear-list;
-  display: flex;
-  align-items: flex-start;
-  flex-wrap: wrap;
-}
-.ingredients__item {
-  width: 100px;
-  min-height: 40px;
-  margin-right: 17px;
-  margin-bottom: 35px;
-}
-.ingredients__counter {
-  width: 54px;
-  margin-top: 10px;
-  margin-left: 36px;
-}
-// diameter
-.diameter__input {
-  margin-right: 8.7%;
-  margin-bottom: 20px;
-  padding-top: 7px;
-  padding-bottom: 6px;
-  cursor: pointer;
-  span {
-    @include r-s16-h19;
-    position: relative;
-    padding-left: 46px;
-    &::before {
-      @include p_center_v;
-      width: 36px;
-      height: 36px;
-      content: "";
-      transition: 0.3s;
-      border-radius: 50%;
-      background-color: $green-100;
-      background-image: url("@/assets/img/diameter.svg");
-      background-repeat: no-repeat;
-      background-position: center;
-    }
-  }
-  &:nth-child(3n) {
-    margin-right: 0;
-  }
-  &--small {
-    span::before {
-      background-size: 18px;
-    }
-  }
-  &--normal {
-    span::before {
-      background-size: 29px;
-    }
-  }
-  &--big {
-    span::before {
-      background-size: 100%;
-    }
-  }
-  &:hover {
-    span::before {
-      box-shadow: $shadow-regular;
-    }
-  }
-  input {
-    &:checked + span::before {
-      box-shadow: $shadow-large;
-    }
-  }
-}
-// counter
-.counter {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.counter__button {
-  $el: &;
-  $size_icon: 50%;
-  position: relative;
-  display: block;
-  width: 16px;
-  height: 16px;
-  margin: 0;
-  padding: 0;
-  cursor: pointer;
-  transition: 0.3s;
-  border: none;
-  border-radius: 50%;
-  outline: none;
-  &--minus {
-    background-color: $purple-100;
-    &::before {
-      @include p_center-all;
-      width: $size_icon;
-      height: 2px;
-      content: "";
-      border-radius: 2px;
-      background-color: $black;
-    }
-    &:hover:not(:active):not(:disabled) {
-      background-color: $purple-200;
-    }
-    &:active:not(:disabled) {
-      background-color: $purple-300;
-    }
-    &:focus:not(:disabled) {
-      box-shadow: $shadow-regular;
-    }
-    &:disabled {
-      cursor: default;
-      &::before {
-        opacity: 0.1;
-      }
-    }
-  }
-  &--plus {
-    background-color: $green-500;
-    &::before {
-      @include p_center-all;
-      width: $size_icon;
-      height: 2px;
-      content: "";
-      border-radius: 2px;
-      background-color: $white;
-    }
-    &::after {
-      @include p_center-all;
-      width: $size_icon;
-      height: 2px;
-      content: "";
-      transform: translate(-50%, -50%) rotate(90deg);
-      border-radius: 2px;
-      background-color: $white;
-    }
-    &:hover:not(:active):not(:disabled) {
-      background-color: $green-400;
-    }
-    &:active:not(:disabled) {
-      background-color: $green-600;
-    }
-    &:focus:not(:disabled) {
-      box-shadow: $shadow-regular;
-    }
-    &:disabled {
-      cursor: default;
-      opacity: 0.3;
-    }
-  }
-  &--orange {
-    background-color: $orange-200;
-    &:hover:not(:active):not(:disabled) {
-      background-color: $orange-100;
-    }
-    &:active:not(:disabled) {
-      background-color: $orange-300;
-    }
-  }
-}
-.counter__input {
-  @include r-s14-h16;
-  box-sizing: border-box;
-  width: 22px;
-  margin: 0;
-  padding: 0 3px;
-  text-align: center;
-  color: $black;
-  border: none;
-  border-radius: 10px;
-  outline: none;
-  background-color: transparent;
-  &:focus {
-    box-shadow: inset $shadow-regular;
-  }
-}
-// visually-hidden
-.visually-hidden {
-  position: absolute;
-  overflow: hidden;
-  clip: rect(0 0 0 0);
-  width: 1px;
-  height: 1px;
-  margin: -1px;
-  padding: 0;
-  white-space: nowrap;
-  border: 0;
-  clip-path: inset(100%);
-}
-// pizza
-.pizza {
-  position: relative;
-  display: block;
-  box-sizing: border-box;
-  width: 100%;
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: 100%;
-  &--foundation--big-creamy {
-    background-image: url("@/assets/img/foundation/big-creamy.svg");
-  }
-  &--foundation--big-tomato {
-    background-image: url("@/assets/img/foundation/big-tomato.svg");
-  }
-  &--foundation--small-creamy {
-    background-image: url("@/assets/img/foundation/small-creamy.svg");
-  }
-  &--foundation--small-tomato {
-    background-image: url("@/assets/img/foundation/small-tomato.svg");
-  }
-}
-.pizza__wrapper {
-  width: 100%;
-  padding-bottom: 100%;
-}
-.pizza__filling {
-  $bl: &;
-  position: absolute;
-  top: 0;
-  left: 0;
-  display: block;
-  width: 100%;
-  height: 100%;
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: 100%;
-  &::before,
-  &::after {
-    display: none;
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    content: "";
-    background-image: inherit;
-  }
-  &--second {
-    &::before {
-      display: block;
-      transform: rotate(45deg);
-    }
-  }
-  &--third {
-    &::before {
-      display: block;
-      transform: rotate(45deg);
-    }
-    &::after {
-      display: block;
-      transform: rotate(-45deg);
-    }
-  }
-  &--ananas,
-  &--ananas.pizza__filling--second::before,
-  &--ananas.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/ananas.svg");
-  }
-  &--bacon,
-  &--bacon.pizza__filling--second::before,
-  &--bacon.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/bacon.svg");
-  }
-  &--blue_cheese,
-  &--blue.pizza__filling--second::before,
-  &--blue.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/blue_cheese.svg");
-  }
-  &--cheddar,
-  &--cheddar.pizza__filling--second::before,
-  &--cheddar.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/cheddar.svg");
-  }
-  &--chile,
-  &--chile.pizza__filling--second::before,
-  &--chile.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/chile.svg");
-  }
-  &--ham,
-  &--ham.pizza__filling--second::before,
-  &--ham.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/ham.svg");
-  }
-  &--jalapeno,
-  &--jalapeno.pizza__filling--second::before,
-  &--jalapeno.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/jalapeno.svg");
-  }
-  &--mozzarella,
-  &--mozzarella.pizza__filling--second::before,
-  &--mozzarella.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/mozzarella.svg");
-  }
-  &--mushrooms,
-  &--mushrooms.pizza__filling--second::before,
-  &--mushrooms.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/mushrooms.svg");
-  }
-  &--olives,
-  &--olives.pizza__filling--second::before,
-  &--olives.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/olives.svg");
-  }
-  &--onion,
-  &--onion.pizza__filling--second::before,
-  &--onion.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/onion.svg");
-  }
-  &--parmesan,
-  &--parmesan.pizza__filling--second::before,
-  &--parmesan.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/parmesan.svg");
-  }
-  &--salami,
-  &---salami.pizza__filling--second::before,
-  &---salami.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/salami.svg");
-  }
-  &--salmon,
-  &--salmon.pizza__filling--second::before,
-  &--salmon.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/salmon.svg");
-  }
-  &--tomatoes,
-  &--tomatoes.pizza__filling--second::before,
-  &--tomatoes.pizza__filling--third::after {
-    background-image: url("@/assets/img/filling-big/tomatoes.svg");
-  }
-}
-// visually-hidden
-.visually-hidden {
-  position: absolute;
-  overflow: hidden;
-  clip: rect(0 0 0 0);
-  width: 1px;
-  height: 1px;
-  margin: -1px;
-  padding: 0;
-  white-space: nowrap;
-  border: 0;
-  clip-path: inset(100%);
-}
-// filling
-.filling {
-  @include r-s14-h16;
-  position: relative;
-  display: block;
-  padding-left: 36px;
-  &::before {
-    @include p_center-v;
-    display: block;
-    width: 32px;
-    height: 32px;
-    content: "";
-    border-radius: 50%;
-    background-color: $white;
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: 80% 80%;
-  }
-  &--tomatoes::before {
-    background-image: url("@/assets/img/filling/tomatoes.svg");
-  }
-  &--ananas::before {
-    background-image: url("@/assets/img/filling/ananas.svg");
-  }
-  &--bacon::before {
-    background-image: url("@/assets/img/filling/bacon.svg");
-  }
-  &--blue_cheese::before {
-    background-image: url("@/assets/img/filling/blue_cheese.svg");
-  }
-  &--cheddar::before {
-    background-image: url("@/assets/img/filling/cheddar.svg");
-  }
-  &--chile::before {
-    background-image: url("@/assets/img/filling/chile.svg");
-  }
-  &--ham::before {
-    background-image: url("@/assets/img/filling/ham.svg");
-  }
-  &--jalapeno::before {
-    background-image: url("@/assets/img/filling/jalapeno.svg");
-  }
-  &--mozzarella::before {
-    background-image: url("@/assets/img/filling/mozzarella.svg");
-  }
-  &--mushrooms::before {
-    background-image: url("@/assets/img/filling/mushrooms.svg");
-  }
-  &--olives::before {
-    background-image: url("@/assets/img/filling/olives.svg");
-  }
-  &--onion::before {
-    background-image: url("@/assets/img/filling/onion.svg");
-  }
-  &--parmesan::before {
-    background-image: url("@/assets/img/filling/parmesan.svg");
-  }
-  &--salami::before {
-    background-image: url("@/assets/img/filling/salami.svg");
-  }
-  &--salmon::before {
-    background-image: url("@/assets/img/filling/salmon.svg");
-  }
-}
+
 </style>
